+++ src/main/kotlin/com/github/xeonkryptos/eclipseprojectcreator/framework/EclipseSupportConfigurable.kt
@@ -1,22 +1,25 @@
 package com.github.xeonkryptos.eclipseprojectcreator.framework
 
 import com.github.xeonkryptos.eclipseprojectcreator.ivy.EclipseIvyUpdater
+import com.github.xeonkryptos.eclipseprojectcreator.util.EclipseUtil
 import com.intellij.facet.FacetTypeRegistry
 import com.intellij.facet.ProjectFacetManager
 import com.intellij.framework.addSupport.FrameworkSupportInModuleConfigurable
-import com.intellij.ide.fileTemplates.FileTemplateUtil
 import com.intellij.openapi.module.Module
-import com.intellij.openapi.module.ModuleUtil
 import com.intellij.openapi.roots.ModifiableModelsProvider
 import com.intellij.openapi.roots.ModifiableRootModel
+import com.intellij.openapi.roots.ModuleRootManager
+import com.intellij.openapi.roots.ModuleRootModel
 import com.intellij.openapi.roots.ui.configuration.libraries.CustomLibraryDescription
-import com.intellij.openapi.vfs.VirtualFileManager
-import java.nio.charset.StandardCharsets
-import java.nio.file.Files
-import java.nio.file.Paths
-import java.nio.file.StandardOpenOption
+import com.intellij.openapi.util.io.FileUtil
+import com.intellij.openapi.vfs.LocalFileSystem
+import java.io.File
+import java.nio.file.Path
 import javax.swing.JComponent
+import org.jdom.output.EclipseJDOMUtil
 import org.jetbrains.idea.eclipse.EclipseXml
+import org.jetbrains.idea.eclipse.conversion.DotProjectFileHelper
+import org.jetbrains.idea.eclipse.conversion.EclipseClasspathWriter
 
 /**
  * @author Xeonkryptos
@@ -37,35 +40,26 @@ class EclipseSupportConfigurable : FrameworkSupportInModuleConfigurable() {
     }
 
     override fun addSupport(module: Module, rootModel: ModifiableRootModel, modifiableModelsProvider: ModifiableModelsProvider) {
-        val moduleBaseDir = Paths.get(ModuleUtil.getModuleDirPath(module))
-        val projectTargetFile = moduleBaseDir.resolve(EclipseXml.PROJECT_FILE)
-        val classpathTargetFile = moduleBaseDir.resolve(EclipseXml.CLASSPATH_FILE)
+        val storageRoot = EclipseUtil.getModuleStorageRoot(module)
+        DotProjectFileHelper.saveDotProjectFile(module, storageRoot)
 
-        val templateAttributes = mapOf(Pair("MODULE_NAME", module.name))
-        Files.newBufferedWriter(projectTargetFile, StandardCharsets.UTF_8, StandardOpenOption.CREATE).use { writer ->
-            this::class.java.getResource("/templates/.project")?.readText(StandardCharsets.UTF_8)?.let { fileContent ->
-                var finalizedProjectFileContent = FileTemplateUtil.mergeTemplate(templateAttributes, fileContent, true)
-                finalizedProjectFileContent = finalizedProjectFileContent.replace("\n", System.lineSeparator())
-                writer.write(finalizedProjectFileContent)
-            }
-        }
-        Files.newBufferedWriter(classpathTargetFile, StandardCharsets.UTF_8, StandardOpenOption.CREATE).use { writer ->
-            this::class.java.getResource("/templates/.classpath")?.readText(StandardCharsets.UTF_8)?.let { fileContent ->
-                val finalizedClasspathFileContent = FileTemplateUtil.mergeTemplate(templateAttributes, fileContent, true)
-                writer.write(finalizedClasspathFileContent)
-            }
-        }
+        val model: ModuleRootModel = ModuleRootManager.getInstance(module)
+        val classpathElement = EclipseClasspathWriter().writeClasspath(model)
 
         FacetTypeRegistry.getInstance().facetTypeIds.firstOrNull { facetTypeId -> facetTypeId.toString() == "IvyIDEA" }.let { ivyIdeaFacetType ->
             if (ivyIdeaFacetType != null && ProjectFacetManager.getInstance(module.project).getModulesWithFacet(ivyIdeaFacetType).contains(module)) {
-                val virtualFileManager = VirtualFileManager.getInstance()
-                virtualFileManager.findFileByNioPath(projectTargetFile)?.let {
-                    EclipseIvyUpdater.updateProjectFileWithIvyNature(module.project, it)
-                }
-                virtualFileManager.findFileByNioPath(classpathTargetFile)?.let {
-                    EclipseIvyUpdater.updateClasspathFileWithIvyContainer(module.project, module, it)
+                EclipseIvyUpdater.updateClasspathFileWithIvyContainer(module, providedClasspathElement = classpathElement)
+
+                LocalFileSystem.getInstance().findFileByNioFile(Path.of(storageRoot, EclipseXml.PROJECT_FILE))?.let { virtualProjectFile ->
+                    EclipseIvyUpdater.updateProjectFileWithIvyNature(module.project, virtualProjectFile)
                 }
             }
+        }
+
+        val classpathFile = File(storageRoot, EclipseXml.CLASSPATH_FILE)
+        if (FileUtil.createIfDoesntExist(classpathFile)) {
+            EclipseJDOMUtil.output(classpathElement, classpathFile, module.project)
+            EclipseUtil.refreshAfterJDOMWrite(classpathFile.path)
         }
     }
 }